Trying to get contours coordinates
detected contours come in the form of MatOfPoints. In order to locate
the detected targets, more efforts are needed. One approach is to use
ImgProc.boundingRect for the mat.
Another approach is to convert mat into list of points and see if we can
find out the maximum and minimum point, etc.