Auto SideLift
Autonomous Command for lifts on the side of the airship. Basic idea is
that robot drives forward for a certain distance (about 14 ft.) to get
right next to the lift, and rotates itself until the target is located
and centered. We’ll definitely have to add in distance adjustments
after centering the targets in the future if we want this to work, but
it should be easy since that’s just what AutoLiftFront does. This
command needs TESTING