Refactor VisionValues
Trying to set it up so that we can just punch in values from the camera, rather than having to do (potentially erroneous) math in our heads. If we can plug in an observed CENTER_X, CENTER_Y, the software should do the rest of the work for us in terms of calculating acceptable error.